pyproject.toml
[project]
name = "mcp-for-postgres-db-demo"
version = "0.1.0"
description = "MCP servers for PostgreSQL with different tool designs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0b1",
    "mcp>=1.3.0",
    "pglast>=7.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "geoalchemy2>=0.15.2",
    "python-dotenv>=1.0.1",
    "azure-identity>=1.16.1",
    "github-copilot-sdk>=0.2.2",
    "agent-framework>=1.0.1",
    "rich>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["servers", "agents", "evals"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0"
]

[tool.ruff]
line-length = 120
target-version = "py312"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]
lint.isort.known-first-party = ["servers", "agents", "evals"]